Finalize move of walkBetweenNodes function into TextSelectionManager
