+++ tests/jest/plugins/plugin.text_selection.test.js
@@ -6,7 +6,6 @@ import {
   genFilter,
   genMap,
   lookAroundWindow,
-  walkBetweenNodes,
   zip,
 } from '@/src/plugins/plugin.text_selection.js';
 
@@ -299,49 +298,6 @@ describe('zip', () => {
 
   test('handles even', () => {
     expect(Array.from(zip([1, 2], [3, 4]))).toEqual([[1, 3], [2, 4]]);
-  });
-});
-
-describe('walkBetweenNodes', () => {
-  const tree = $(FAKE_XML_MULT_LINES);
-
-  test('handles empty', () => {
-    const result = Array.from(walkBetweenNodes(tree[0], tree[0]));
-    expect(result).toHaveLength(1);
-    expect(result[0]).toBe(tree[0]);
-  });
-
-  test('Words on same line', () => {
-    const start = tree.find('WORD')[2];
-    const end = start.nextElementSibling;
-    // Use first child so we hit the text nodes
-    const result = Array.from(walkBetweenNodes(start.firstChild, end.firstChild));
-    expect(result).toHaveLength(5);
-    expect(result[0].nodeType).toBe(Node.TEXT_NODE);
-    expect(result[0].textContent).toBe('false ');
-    expect(result[1]).toBe(start);
-    // Whitespace
-    expect(result[2].nodeType).toBe(Node.TEXT_NODE);
-    expect(result[2].textContent).toMatch(/^\s*$/);
-    expect(result[3]).toBe(end);
-    expect(result[4].nodeType).toBe(Node.TEXT_NODE);
-    expect(result[4].textContent).toBe('judgment ');
-  });
-
-  test('Words on different lines', () => {
-    const start = tree.find('WORD')[2];
-    const end = tree.find('WORD')[19];
-    const result = Array.from(walkBetweenNodes(start.firstChild, end.firstChild));
-    // Expect two LINES in result
-    expect(result.filter(x => x.nodeName == 'LINE')).toHaveLength(2);
-    // Expect 18 WORDs in result
-    expect(result.filter(x => x.nodeName == 'WORD')).toHaveLength(18);
-    // First word should be the start word
-    expect(result[0].parentElement).toBe(start);
-    expect(result[0].textContent).toBe('false ');
-    // Last word should be the end word
-    expect(result[result.length - 1].parentElement).toBe(end);
-    expect(result[result.length - 1].textContent).toBe('Suppose');
   });
 });
